requirements.txt
RPi.GPIO
gpiozero
loguru
